+++ src/main/java/net/sf/jabref/gui/CleanUpAction.java
@@ -31,6 +31,7 @@ import com.jgoodies.forms.factories.Borders;
 
 import net.sf.jabref.*;
 import net.sf.jabref.external.ExternalFileType;
+import net.sf.jabref.gui.preftabs.ImportSettingsTab;
 import net.sf.jabref.imports.HTMLConverter;
 import net.sf.jabref.imports.CaseKeeper;
 import net.sf.jabref.imports.UnitFormatter;
