+++ src/main/java/net/sf/jabref/gui/preftabs/ImportSettingsTab.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref;
+package net.sf.jabref.gui.preftabs;
 
 import java.awt.BorderLayout;
 import java.awt.event.ActionEvent;
@@ -21,7 +21,7 @@ import java.awt.event.ActionListener;
 
 import javax.swing.*;
 
-import net.sf.jabref.gui.preftabs.PrefsTab;
+import net.sf.jabref.Globals;
 import spl.gui.ImportDialog;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
