+++ src/main/java/net/sf/jabref/util/Util.java
@@ -87,7 +87,7 @@ import net.sf.jabref.EasyDateFormat;
 import net.sf.jabref.EntryMarker;
 import net.sf.jabref.GUIGlobals;
 import net.sf.jabref.Globals;
-import net.sf.jabref.ImportSettingsTab;
+import net.sf.jabref.gui.preftabs.ImportSettingsTab;
 import net.sf.jabref.JabRefFrame;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.MetaData;
